[cloud/client] Fix compilation warnings
Always use 64-bit integers on file sizes, even with 32-bit compilation

+++ src/libraries/client/Context.h
@@ -85,7 +85,7 @@ public slots:
     void OnNetworkRequest( HttpCommand cmd, const QNetworkRequest& req );
     void OnGetSession();
     void OnCloseDownload( size_t id, bool valid );
-    void OnDownloadProgress( size_t id, size_t current, int progress );
+    void OnDownloadProgress( size_t id, int64_t current, int progress );
     void OnDownloadError( size_t id, const QString& error );
     void ParsePackages();
     void OnSslErrors( QNetworkReply*, const QList< QSslError >& errors );
